Create process per connection
This code allows the HTTP server to handle multiple connections simultanteously.
It does this by creating a process for each connection. This clearly isn't the
ideal solution as creating a process is pretty expensive, but it's a good simple
solution that provides connection isolation.

In the future I may implement threaded connections, or possibly even some kind
of event based system.